backend/requirements.txt
# Backend Requirements - Resolved Versions for Compatibility
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloc==0.2.5
colored==2.2.4
cron-descriptor==1.4.5
cryptography==42.0.8
daphne==4.0.0
defusedxml==0.8.0rc2
Django==5.1.1
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-elasticsearch-dsl==8.0
django-environ==0.11.2
django-extensions==3.2.3
django-gravatar==0.1.0
django-prometheus==2.3.1
django-redis==5.4.0
django-templated-mail==1.1.1
django-timezone-field==7.0
django_csp==3.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.26.2
drf-yasg==1.21.7
elastic-transport==8.13.1
elasticsearch>=8.0.0,<9.0.0  # Relaxed for compatibility
elasticsearch-dsl>=8.0.0,<9.0.0
gunicorn==22.0.0
idna==3.8
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kafka-python==2.0.2
kombu==5.4.0
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.1
Pillow==9.5.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
pytest==8.3.3
pytest-django==4.8.0
pytest-asyncio==0.24.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
requests>=2.31.0,<3.0.0  # Ensure compatibility with urllib3
requests-oauthlib==2.0.0
rpds-py==0.19.0
sentry-sdk>=2.14.0,<3.0  # Relaxed to work with urllib3 versions
setuptools==74.1.2
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse>=0.4.4,<0.5  # Adjusted for compatibility
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3>=1.26.11,<3.0  # Relaxed to meet requirements from multiple libraries
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0

# Removed Djongo and added MongoEngine
mongoengine==0.27.0  # MongoEngine for ODM
pymongo>=4.0,<5.0  # PyMongo for MongoEngine and direct access to MongoDB
